fix: 去掉 drowdown 定位
- 外部变化引起定位属性不对
- 把类型声明迁移到 interface 文件
